% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coefficient.R
\name{Beta.hat}
\alias{Beta.hat}
\title{Estimate the coefficient vector β}
\usage{
Beta.hat(X, Y, method = "BFGS")
}
\arguments{
\item{X}{An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors(Not including intercept).}

\item{Y}{A factor vector indicating the category of response}

\item{method}{The method to be used by optim() function include "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent". The default value is "BFGS"}
}
\value{
The optimal coefficient vector β
}
\description{
{Estimate the coefficient vector β which includes the independent variables/predictors plus the intercept}
}
\author{
Fangjian Yang
}
