% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceInterval.R
\name{boot.confi}
\alias{boot.confi}
\title{Calculate confidence intervals by bootstrap}
\usage{
boot.confi(X, Y, alpha, B = 20)
}
\arguments{
\item{X}{An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors(Not including intercept).}

\item{Y}{A factor vector indicating the category of response}

\item{alpha}{A variable that indicating the significance level}

\item{B}{A variable vector indicating the number of bootstraps(Default 20)}
}
\value{
Bootstrap Confidence intervals
}
\description{
{Bootstrap Confidence intervals: the user must be able to choose (i) the significance level α  to obtain for the 1−α  confidence intervals for β , and (ii) the number of bootstraps which by default will be 20.}
}
\author{
Fangjian Yang
}
