% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfusionMatrix.R
\name{confusion.matrix}
\alias{confusion.matrix}
\title{Calculate the confusion matrix and required metrics.}
\usage{
confusion.matrix(pred.value, actual.value, cutoff = 0.5)
}
\arguments{
\item{pred.value}{The predicted value of fitted logistic model}

\item{actual.value}{The actual value of the data set}

\item{cutoff}{The cut-off values for measuring metrics.}
}
\value{
The confusion matrix and six metrics
}
\description{
{This function calculates the confusion matrix by comparing the predicted values with the true values and calculates the following six metrics: 1.Prevalence, 2.Accuracy, 3.Sensitivity, 4.Specificity, 5.False.Discovery.Rate, 6.Diagnostic.Odds.Ratio. The size of the cut-off value may affect the output results and the performance of the model. Therefore, the cut-off value of this function is user-definable, and its default value is 0.5. }
}
\author{
Leshan Zhao
}
