% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BascFunction.R
\name{logistic.regression}
\alias{logistic.regression}
\title{Summary of the output of the logistic regression function}
\usage{
logistic.regression(X, Y, method = "BFGS", cutoff = 0.5, alpha = 0.1, B = 20)
}
\arguments{
\item{X}{X An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors.}

\item{Y}{A factor vector indicating the category of response}

\item{method}{The method to be used by optim() function include "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent". The default value is "BFGS"}

\item{cutoff}{The cut-off value for measuring metrics.}

\item{alpha}{A variable that indicating the significance level}

\item{B}{A variable indicating the number of bootstraps(Default 20)}
}
\value{
The list of all output
}
\description{
{Inputting the data set into this function can fit a logistic regression, and can get the following output: 1. estimated logistic regression model; 2. confidence intervals for the parameters in the model; 3. logistic regression curves for each predictor; 4. confusion matrix; 5. output for the six metrics; and 6. metrics with different cutoff values.}
}
\author{
Fangjian Yang
}
