% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticCurve.R
\name{logistic_plot}
\alias{logistic_plot}
\title{Plot of the fitted logistic curve to the actual values}
\usage{
logistic_plot(X, Y)
}
\arguments{
\item{X}{X An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors.}

\item{Y}{A factor vector indicating the category of response}
}
\value{
The list of all logistic curve of predictor
}
\description{
{This function plots the input X and Y through the ggplot function for logistic regression curves, and if the input dataset contains multiple predictors, the logistic curve is plotted separately for each predictor and all images are combined into a list data.}
}
\author{
Fangjian Yang
}
