% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfusionMatrix.R
\name{logistic_pred}
\alias{logistic_pred}
\title{Calculate the predicted value.}
\usage{
logistic_pred(model, X)
}
\arguments{
\item{model}{The parameters of the logistic regression model.}

\item{X}{X An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors.}
}
\value{
The confusion matrix and six metrics
}
\description{
{Input the parameters of the logistic regression model and predictor variables to obtain the predicted values.}
}
\author{
Fangjian Yang
}
