% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BascFunction.R
\name{loss_func}
\alias{loss_func}
\title{The numerical optimization loss function}
\usage{
loss_func(beta, y, xi)
}
\arguments{
\item{beta}{The current beta vector(Including intercept).}

\item{y}{A training factor vector indicating the category of response}

\item{xi}{An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors(Including intercept).}
}
\value{
The estimator of loss function
}
\description{
{The estimator to be computed using numerical optimization is the following: \deqn{\hat\beta=\underset{\beta}{argmin}\sum_{i=1}^{n}（-y_{i}\cdot ln(p_{i})-(1-y_{i} \cdot ln(1-p_{i}))）}\\}
}
\author{
Fangjian Yang
}
