% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cut-offValue.R
\name{metrics.table}
\alias{metrics.table}
\title{Calculate the six metrics of the model with different cut-off values}
\usage{
metrics.table(X, Y, interval = c(0.1, 0.9), step = 0.1)
}
\arguments{
\item{X}{X An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors.}

\item{Y}{A factor vector indicating the category of response}

\item{interval}{The range of cut-off value (Default value is from 0.1 to 0.9).}

\item{step}{Step length of the interval of cut-off values(Default value is 0.1).}
}
\value{
The dataframe of six metrics with different cutoff value
}
\description{
{This function allows the user to plot the evaluation of the six metrics on a grid of predicted cutoff values from 0.1 to 0.9, in steps of 0.1.}
}
\author{
Leshan Zhao
}
