% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BascFunction.R
\name{P.i}
\alias{P.i}
\title{Calculate the Pi vector for loss function}
\usage{
P.i(beta, Xi)
}
\arguments{
\item{beta}{The current beta vector(Including intercept).}

\item{Xi}{An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors(Including intercept).}
}
\value{
The Pi value for loss function
}
\description{
{the Pi vector for loss function from the formula \deqn{P_{i}=\frac{1}{1+e^{-x^{T}_{i}\beta}}}\\}
}
\author{
Fangjian Yang
}
