% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BascFunction.R
\name{X.format}
\alias{X.format}
\title{Uniform data format for predictor variables}
\usage{
X.format(X, intercept = F)
}
\arguments{
\item{X}{An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors.}

\item{intercept}{A logical variable indicating whether to add an intercept to the predictor variables}
}
\value{
A matrix containing the predictor variables
}
\description{
This function converts the input predictor variables into matrix format and adds intercepts variable accord by user
}
\author{
Fangjian Yang
}
